Rename load.py to cohort.py
